Remove unmodified Vanilla files
Some files in the mod are the same as the Vanilla.
They are useless, we remove them in order to be more clear, and mostly
more compatible with other mods.

Nota: There are many province files in this case. Like there is more
than 1000 files in the directory, I cannot remove them quickly with git.
So, we have not removed all, we need to remove other later