Fix bug with rex vassal counting
The event RIRPR.930 was bad written. In the previous way, all the flag
were set by only one call. It was be better to write on the other way:
check there is already 4 flags, then check there is already 3 flags ...
I have written it again with a variable because it is a better and
simpler way to do this.

Fixing this bug create a bug in case there is less than 5 king which are
vassals of the Pope. So in this case, the flag verip_pr_rex is clear in
order to call Nouveau_pape_verif later