Add a decision to go to Canossa #4
The Emperor is not obliged, like historically, to wait 5 years to go to
Canossa.
So I have created a decision and indicate its existence when the Emperor
is excommunicated.
I have keep the events which called every 5 years the Emperor to go to
Canossa.

The new decision is restricted to the player.