Fix bug with king vassal counting #4
1795da0 is bugged.
With parallelism problems, the variable is equal to 0 or 1 (all the
vassal increase by 1 the variable at the same moment, so the variable is
equal to ... 1)

To fix this issue, I use a simple count in a scope.
In addition, we don't need anymore the RIRPR.930 because there is no
more to do with it.
The RIRPR.929 is still used to avert the vassals.