Renova imperi is a victory of investiture controversy #4
Faras explains in this video (https://youtu.be/9bM-l8FzTMw?t=19m8s : I
have set at 19m8s for the interesting moment) that the Renova Imperi is
a victory if we lost the investiture controversy.
But it was not implemented.
So I have implemented it by given the flag of the investiture
controversy and add 2 additional check of this flag.